feat: extract gateway_comments_rest and mount via capability registry
Move the /comments/{id} and /posts/{id}/comments REST routes into a new
gateway_comments_rest capability package (mirrors gateway_posts_rest),
consuming gateway_comments.service.CommentsService via a shared
get_comments_service dependency. gateway_oss/api/v1/comments.py and
posts.py are deleted and unregistered from api/v1/__init__.py now that
both routes are served exactly once through the capability registry.
Also add the missing gateway_comments/__init__.py for sibling-package
consistency with the other domain packages.
