+++ packages/gateway_comments_rest/pyproject.toml
@@ -1,0 +1,21 @@
+[project]
+name = "gateway_comments_rest"
+version = "0.1.0"
+requires-python = ">=3.11"
+dependencies = ["gateway_core", "gateway_rest_common", "gateway_comments", "gateway_notes", "fastapi>=0.138.2"]
+
+[project.entry-points."substack_gateway.capabilities"]
+comments_rest = "gateway_comments_rest:capability"
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/gateway_comments_rest"]
+
+[tool.uv.sources]
+gateway_core = { workspace = true }
+gateway_rest_common = { workspace = true }
+gateway_comments = { workspace = true }
+gateway_notes = { workspace = true }
